news/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
black==24.10.0
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
inflection==0.5.1
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.11.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
